backend/requirements.txt
fastapi
uvicorn
ultralytics
scikit-learn
pandas
numpy
joblib
pillow
python-multipart
